# packages/api/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Integration Hub API",
    "description": "REST API for the Integration Hub healthcare device interoperability platform. Connects laboratory analyzers, imaging modalities, and medical devices to LIS/HIS/EMR/RIS/PACS systems.\n\n## Authentication\n\nAll `/api/v1` routes (except `/api/v1/health`) require `Authorization: Bearer <key>` header. Keys are created via `POST /api/v1/keys` (admin only). Roles: `admin`, `engineer`, `operator`, `viewer`.\n\n## Versioning\n\nThis is API version 1 (`/api/v1/`). Breaking changes will introduce a new version (`/api/v2/`) with both versions supported during a transition period. Non-breaking additions (new fields, new optional parameters, new endpoints) are added to v1 without notice.\n\n## Rate Limits\n\nNo rate limits enforced in v1. Cloud deployments may add per-tenant quotas in later phases.",
    "version": "1.0.0",
    "contact": {
      "name": "Integration Hub"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/fleet/gateways": {
      "post": {
        "tags": [
          "Fleet"
        ],
        "summary": "Register an edge gateway (returns the pairing code exactly once)",
        "responses": {
          "201": {
            "description": "Gateway registered"
          }
        },
        "operationId": "registerGateway",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "facilityId"
                ],
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "facilityId": {
                    "type": "string"
                  },
                  "orgId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Fleet"
        ],
        "summary": "List fleet gateways (pairing codes never listed)",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "operationId": "listGateways"
      }
    },
    "/api/v1/fleet/gateways/{id}": {
      "get": {
        "tags": [
          "Fleet"
        ],
        "summary": "Get one gateway",
        "responses": {
          "200": {
            "description": "Gateway"
          },
          "404": {
            "description": "Not found"
          }
        },
        "operationId": "getGateway",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/fleet/gateways/{id}/revoke": {
      "post": {
        "tags": [
          "Fleet"
        ],
        "summary": "Revoke a gateway (its sync credential stops working)",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "operationId": "revokeGateway",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/facilities": {
      "get": {
        "tags": [
          "Fleet"
        ],
        "summary": "List facilities in the org",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "operationId": "listFacilities"
      },
      "post": {
        "tags": [
          "Fleet"
        ],
        "summary": "Create a facility",
        "responses": {
          "201": {
            "description": "Facility created"
          }
        },
        "operationId": "createFacility",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "slug": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/provision/claim": {
      "post": {
        "tags": [
          "Fleet"
        ],
        "summary": "Claim a pairing code (public; returns the provisioning bundle with the gateway API key exactly once)",
        "responses": {
          "201": {
            "description": "Provisioning bundle"
          },
          "409": {
            "description": "Code consumed or wrong state"
          },
          "410": {
            "description": "Pairing code expired"
          }
        },
        "operationId": "claimProvisioning",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "pairingCode"
                ],
                "properties": {
                  "pairingCode": {
                    "type": "string"
                  },
                  "edgeVersion": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/ingest": {
      "post": {
        "tags": [
          "Fleet"
        ],
        "summary": "D11 sync ingest (gateway credential: x-hub-gateway + Bearer)",
        "responses": {
          "200": {
            "description": "Applied; returns appliedThrough + accepted"
          },
          "401": {
            "description": "Invalid gateway credential"
          }
        },
        "operationId": "syncIngest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "entries"
                ],
                "properties": {
                  "entries": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "seq",
                        "table",
                        "op",
                        "pk",
                        "payload",
                        "createdAt"
                      ],
                      "properties": {
                        "seq": {
                          "type": "integer"
                        },
                        "table": {
                          "type": "string",
                          "enum": [
                            "messages",
                            "devices"
                          ]
                        },
                        "op": {
                          "type": "string",
                          "enum": [
                            "INSERT",
                            "UPDATE"
                          ]
                        },
                        "pk": {
                          "type": "string"
                        },
                        "payload": {},
                        "facilityId": {
                          "type": "string"
                        },
                        "orgId": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/platform/flags/{key}": {
      "put": {
        "tags": [
          "Fleet"
        ],
        "summary": "Set a feature flag",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "operationId": "setFlag",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "delete": {
        "tags": [
          "Fleet"
        ],
        "summary": "Remove a feature flag",
        "responses": {
          "204": {
            "description": "Removed"
          }
        },
        "operationId": "removeFlag",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/platform/quotas/{facilityId}": {
      "put": {
        "tags": [
          "Fleet"
        ],
        "summary": "Set a facility quota",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "operationId": "setQuota",
        "parameters": [
          {
            "name": "facilityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "delete": {
        "tags": [
          "Fleet"
        ],
        "summary": "Remove a facility quota",
        "responses": {
          "204": {
            "description": "Removed"
          }
        },
        "operationId": "removeQuota",
        "parameters": [
          {
            "name": "facilityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/licenses/{key}": {
      "put": {
        "tags": [
          "Fleet"
        ],
        "summary": "Upsert a license",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "operationId": "putLicense",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "delete": {
        "tags": [
          "Fleet"
        ],
        "summary": "Remove a license",
        "responses": {
          "204": {
            "description": "Removed"
          }
        },
        "operationId": "removeLicense",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/licenses/{facilityId}/entitlement": {
      "get": {
        "tags": [
          "Fleet"
        ],
        "summary": "Entitlement evaluation for a facility (tier + feature gates)",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "operationId": "getEntitlement",
        "parameters": [
          {
            "name": "facilityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key secret (created via POST /api/v1/keys). Format: ihk_..."
      }
    },
...
  }
}